Refactor buffer and scrolling management.
The content of the buffer is now stored into a char *data field and all of the
Text stuff has been removed. This made dealing with the content simpler and
also allowed a finally decent line-based scrolling implementation which also
survives to resize.

This is a quick and dirty (but nicely tested) commit which also fix a few
corner cases. I have some refactor and few code cleanups in my todo list.