Better server messages handling.
A new type has been defined to bind IRC commands to callbacks:

typedef struct {
	char *name;
	void (*func)(char *, char *, char *);
} Message;

Interface is not much elegant but do its job well and allows each callback to
do its own parsing. Thus parsesrv() is arranged to find the appropriate
callback for the executing command and, if found, run it. All unhandled
commands are just discarded.

This commit also removes all the draw(), drawbuf() and drawcmdln() call from
all around the code and replace them with a need_redraw field of the buffer.
This is a raw implementation and distinction between buffer and command line
has to be implemented.

All of this needs to be deeply tested.