Handle non-standard JOIN in recv_join() and few code cleanups.
Some server like irc.oftc.net puts a ':' character before the channel, thus it
makes the parser think that there is an additional parameter. Now recv_join()
try to get the channel name from its proper place but if it's not filled, the
`txt` variable is used instead (which contains the value after the ':').