Use globals to store I/O buffers.
Lost connections are now handled in main(). This makes sense because the
parsesrv() function should be called only if there is data available from the
server, else we should just disconnects. Which is how it works now.

When a connection goes down, the message "Remote host closed connection" is
sent to all buffers.