Initial effort to support UTF-8.
Implementation is based on two macros which works on raw bytes (char *):
UTF8BYTES() which returns the size (total bytes) of a code point, and
UTF8CBYTE() which test for a continuation byte. This may not be the perfect
solution but seems to be working for all tested scenarios. Currently wchar_t is
only used to compute columns width but this may change in future.

In future I may cosider switching to libgrapheme or the like but, even if there
are various cleanups needed and some refactoring, since it works surprisingly
well, I will stuck with this simple design for a while.