Fix buffer management in cmdln_submit().
The old implementation was ignoring the fact that parsecmd() may cause sel to
be changed or even destroyed. Since there's no way to know if the current
buffer has been free()ed before accessing the associated data, we just copy the
data and pass it around to parsecmd() which cannot interfere with the buffer
anymore. Finally, we free() the data and we're done.